+++ moment-timezone-utils.js
@@ -1,5 +1,5 @@
 //! moment-timezone-utils.js
-//! version : 0.2.5
+//! version : 0.3.0
 //! author : Tim Wood
 //! license : MIT
 //! github.com/moment/moment-timezone